requirements.txt
asyncio-mqtt==0.16.2
aiohttp==3.9.5
aioredis==2.0.1
websockets==12.0

# Container and orchestration
docker==7.1.0
kubernetes==29.0.0

# Network and system monitoring
netifaces==0.11.0
psutil==5.9.8
speedtest-cli==2.1.3

# Data processing and storage
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
alembic==1.13.1

# Security and encryption
cryptography==42.0.7
pyjwt==2.8.0

# Configuration and serialization
pyyaml==6.0.1
pydantic==2.7.1

# Logging and monitoring
structlog==24.1.0
prometheus-client==0.20.0

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0

# Development tools
black==24.4.2
flake8==7.0.0
mypy==1.10.0
pre-commit==3.7.1

prometheus-client==0.20.0
matplotlib==3.8.4
seaborn==0.13.2
pandas==2.2.2
numpy==1.26.4